// utils/generated/subgraph.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export interface Scalars {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  BigDecimal: any;
  BigInt: any;
  Bytes: any;
}






export interface Block_Height {
  readonly hash?: Maybe<Scalars['Bytes']>;
  readonly number?: Maybe<Scalars['Int']>;
}

export interface Bundle {
  readonly __typename?: 'Bundle';
  readonly id: Scalars['ID'];
  /** BNB price, in USD */
  readonly bnbPrice: Scalars['BigDecimal'];
}
// ...
export interface Burn {
  readonly __typename?: 'Burn';
  readonly id: Scalars['ID'];
  readonly transaction: Transaction;
  readonly timestamp: Scalars['BigInt'];
  readonly pair: Pair;
  readonly liquidity: Scalars['BigDecimal'];
  readonly sender?: Maybe<Scalars['Bytes']>;
  readonly amount0?: Maybe<Scalars['BigDecimal']>;
  readonly amount1?: Maybe<Scalars['BigDecimal']>;
  readonly to?: Maybe<Scalars['Bytes']>;
  readonly logIndex?: Maybe<Scalars['BigInt']>;
  readonly amountUSD?: Maybe<Scalars['BigDecimal']>;
  readonly needsComplete: Scalars['Boolean'];
  readonly feeTo?: Maybe<Scalars['Bytes']>;
  readonly feeLiquidity?: Maybe<Scalars['BigDecimal']>;
}
// ...
export interface Factory {
  readonly __typename?: 'Factory';
  readonly id: Scalars['ID'];
  /** Total of pairs */
  readonly totalPairs: Scalars['BigInt'];
  /** Total of transactions */
  readonly totalTransactions: Scalars['BigInt'];
  readonly totalVolumeUSD: Scalars['BigDecimal'];
  readonly totalVolumeBNB: Scalars['BigDecimal'];
  readonly untrackedVolumeUSD: Scalars['BigDecimal'];
  readonly totalLiquidityUSD: Scalars['BigDecimal'];
  readonly totalLiquidityBNB: Scalars['BigDecimal'];
}
// ...
export interface Mint {
  readonly __typename?: 'Mint';
  readonly id: Scalars['ID'];
  readonly transaction: Transaction;
  readonly timestamp: Scalars['BigInt'];
  readonly pair: Pair;
  readonly to: Scalars['Bytes'];
  readonly liquidity: Scalars['BigDecimal'];
  readonly sender?: Maybe<Scalars['Bytes']>;
  readonly amount0?: Maybe<Scalars['BigDecimal']>;
  readonly amount1?: Maybe<Scalars['BigDecimal']>;
  readonly logIndex?: Maybe<Scalars['BigInt']>;
  readonly amountUSD?: Maybe<Scalars['BigDecimal']>;
  readonly feeTo?: Maybe<Scalars['Bytes']>;
  readonly feeLiquidity?: Maybe<Scalars['BigDecimal']>;
}
// ...
export interface Pair {
  readonly __typename?: 'Pair';
  readonly id: Scalars['ID'];
  readonly name: Scalars['String'];
  readonly token0: Token;
  readonly token1: Token;
  readonly reserve0: Scalars['BigDecimal'];
  readonly reserve1: Scalars['BigDecimal'];
  readonly totalSupply: Scalars['BigDecimal'];
  readonly reserveBNB: Scalars['BigDecimal'];
  readonly reserveUSD: Scalars['BigDecimal'];
  readonly trackedReserveBNB: Scalars['BigDecimal'];
  readonly token0Price: Scalars['BigDecimal'];
  readonly token1Price: Scalars['BigDecimal'];
  readonly volumeToken0: Scalars['BigDecimal'];
  readonly volumeToken1: Scalars['BigDecimal'];
  readonly volumeUSD: Scalars['BigDecimal'];
  readonly untrackedVolumeUSD: Scalars['BigDecimal'];
  readonly totalTransactions: Scalars['BigInt'];
  readonly block: Scalars['BigInt'];
  readonly timestamp: Scalars['BigInt'];
  readonly pairHourData: ReadonlyArray<PairHourData>;
  readonly mints: ReadonlyArray<Mint>;
  readonly burns: ReadonlyArray<Burn>;
  readonly swaps: ReadonlyArray<Swap>;
}
// ...
export interface PairDayData {
  readonly __typename?: 'PairDayData';
  readonly id: Scalars['ID'];
  readonly date: Scalars['Int'];
  readonly pairAddress: Scalars['Bytes'];
  readonly token0: Token;
  readonly token1: Token;
  readonly reserve0: Scalars['BigDecimal'];
  readonly reserve1: Scalars['BigDecimal'];
  readonly totalSupply: Scalars['BigDecimal'];
  readonly reserveUSD: Scalars['BigDecimal'];
  readonly dailyVolumeToken0: Scalars['BigDecimal'];
  readonly dailyVolumeToken1: Scalars['BigDecimal'];
  readonly dailyVolumeUSD: Scalars['BigDecimal'];
  readonly dailyTxns: Scalars['BigInt'];
}
// ...
export interface PairHourData {
  readonly __typename?: 'PairHourData';
  readonly id: Scalars['ID'];
  readonly hourStartUnix: Scalars['Int'];
  readonly pair: Pair;
  readonly reserve0: Scalars['BigDecimal'];
  readonly reserve1: Scalars['BigDecimal'];
  readonly totalSupply: Scalars['BigDecimal'];
  readonly reserveUSD: Scalars['BigDecimal'];
  readonly hourlyVolumeToken0: Scalars['BigDecimal'];
  readonly hourlyVolumeToken1: Scalars['BigDecimal'];
  readonly hourlyVolumeUSD: Scalars['BigDecimal'];
  readonly hourlyTxns: Scalars['BigInt'];
}
// ...
export interface PlatformDayData {
  readonly __typename?: 'PlatformDayData';
  readonly id: Scalars['ID'];
  readonly date: Scalars['Int'];
  readonly dailyVolumeBNB: Scalars['BigDecimal'];
  readonly dailyVolumeUSD: Scalars['BigDecimal'];
  readonly dailyVolumeUntracked: Scalars['BigDecimal'];
  readonly totalVolumeBNB: Scalars['BigDecimal'];
  readonly totalLiquidityBNB: Scalars['BigDecimal'];
  readonly totalVolumeUSD: Scalars['BigDecimal'];
  readonly totalLiquidityUSD: Scalars['BigDecimal'];
  readonly totalTransactions: Scalars['BigInt'];
}
// ...
export interface Query {
  readonly __typename?: 'Query';
  readonly factory?: Maybe<Factory>;
  readonly factories: ReadonlyArray<Factory>;
  readonly bundle?: Maybe<Bundle>;
  readonly bundles: ReadonlyArray<Bundle>;
  readonly token?: Maybe<Token>;
  readonly tokens: ReadonlyArray<Token>;
  readonly pair?: Maybe<Pair>;
  readonly pairs: ReadonlyArray<Pair>;
  readonly transaction?: Maybe<Transaction>;
  readonly transactions: ReadonlyArray<Transaction>;
  readonly mint?: Maybe<Mint>;
  readonly mints: ReadonlyArray<Mint>;
  readonly burn?: Maybe<Burn>;
  readonly burns: ReadonlyArray<Burn>;
  readonly swap?: Maybe<Swap>;
  readonly swaps: ReadonlyArray<Swap>;
  readonly platformDayData?: Maybe<PlatformDayData>;
  readonly platformDayDatas: ReadonlyArray<PlatformDayData>;
  readonly pairHourData?: Maybe<PairHourData>;
  readonly pairHourDatas: ReadonlyArray<PairHourData>;
  readonly pairDayData?: Maybe<PairDayData>;
  readonly pairDayDatas: ReadonlyArray<PairDayData>;
  readonly tokenDayData?: Maybe<TokenDayData>;
  readonly tokenDayDatas: ReadonlyArray<TokenDayData>;
  /** Access to subgraph metadata */
  readonly _meta?: Maybe<_Meta_>;
}
// ...
export interface Swap {
  readonly __typename?: 'Swap';
  readonly id: Scalars['ID'];
  readonly transaction: Transaction;
  readonly timestamp: Scalars['BigInt'];
  readonly pair: Pair;
  readonly sender: Scalars['Bytes'];
  readonly from: Scalars['Bytes'];
  readonly amount0In: Scalars['BigDecimal'];
  readonly amount1In: Scalars['BigDecimal'];
  readonly amount0Out: Scalars['BigDecimal'];
  readonly amount1Out: Scalars['BigDecimal'];
  readonly to: Scalars['Bytes'];
  readonly logIndex?: Maybe<Scalars['BigInt']>;
  readonly amountUSD: Scalars['BigDecimal'];
}
// ...
export interface Token {
  readonly __typename?: 'Token';
  readonly id: Scalars['ID'];
  /** Name */
  readonly name: Scalars['String'];
  /** Symbol */
  readonly symbol: Scalars['String'];
  /** Decimals */
  readonly decimals: Scalars['BigInt'];
  readonly tradeVolume: Scalars['BigDecimal'];
  readonly tradeVolumeUSD: Scalars['BigDecimal'];
  readonly untrackedVolumeUSD: Scalars['BigDecimal'];
  readonly totalTransactions: Scalars['BigInt'];
  readonly totalLiquidity: Scalars['BigDecimal'];
  readonly derivedBNB?: Maybe<Scalars['BigDecimal']>;
  readonly derivedUSD?: Maybe<Scalars['BigDecimal']>;
  readonly tokenDayData: ReadonlyArray<TokenDayData>;
  readonly pairDayDataBase: ReadonlyArray<PairDayData>;
  readonly pairDayDataQuote: ReadonlyArray<PairDayData>;
  readonly pairBase: ReadonlyArray<Pair>;
  readonly pairQuote: ReadonlyArray<Pair>;
}
// ...
export interface TokenDayData {
  readonly __typename?: 'TokenDayData';
  readonly id: Scalars['ID'];
  readonly date: Scalars['Int'];
  readonly token: Token;
  readonly dailyVolumeToken: Scalars['BigDecimal'];
  readonly dailyVolumeBNB: Scalars['BigDecimal'];
  readonly dailyVolumeUSD: Scalars['BigDecimal'];
  readonly dailyTxns: Scalars['BigInt'];
  readonly totalLiquidityToken: Scalars['BigDecimal'];
  readonly totalLiquidityBNB: Scalars['BigDecimal'];
  readonly totalLiquidityUSD: Scalars['BigDecimal'];
  readonly priceUSD: Scalars['BigDecimal'];
}
// ...
export interface Transaction {
  readonly __typename?: 'Transaction';
  readonly id: Scalars['ID'];
  readonly block: Scalars['BigInt'];
  readonly timestamp: Scalars['BigInt'];
  readonly mints: ReadonlyArray<Maybe<Mint>>;
  readonly burns: ReadonlyArray<Maybe<Burn>>;
  readonly swaps: ReadonlyArray<Maybe<Swap>>;
}
// ...
export interface _Meta_ {
  readonly __typename?: '_Meta_';
  /**
   * Information about a specific subgraph block. The hash of the block
   * will be null if the _meta field has a block constraint that asks for
   * a block number. It will be filled if the _meta field has no block constraint
   * and therefore asks for the latest  block
   */
  readonly block: _Block_;
  /** The deployment ID */
  readonly deployment: Scalars['String'];
  /** If `true`, the subgraph encountered indexing errors at some past block */
  readonly hasIndexingErrors: Scalars['Boolean'];
}
// ...
export type FactoriesQueryVariables = Exact<{
  blockNumber: Scalars['Int'];
  factoryAddress: Scalars['ID'];
}>;


export type FactoriesQuery = (
  { readonly __typename?: 'Query' }
  & { readonly factories: ReadonlyArray<(
    { readonly __typename?: 'Factory' }
    & Pick<Factory, 'id' | 'totalVolumeUSD' | 'totalVolumeBNB' | 'untrackedVolumeUSD' | 'totalLiquidityUSD' | 'totalLiquidityBNB' | 'totalTransactions' | 'totalPairs'>
  )> }
);

export type CurrentFactoriesQueryVariables = Exact<{
  factoryAddress: Scalars['ID'];
}>;


export type CurrentFactoriesQuery = (
  { readonly __typename?: 'Query' }
  & { readonly factories: ReadonlyArray<(
    { readonly __typename?: 'Factory' }
    & Pick<Factory, 'id' | 'totalVolumeUSD' | 'totalVolumeBNB' | 'untrackedVolumeUSD' | 'totalLiquidityUSD' | 'totalLiquidityBNB' | 'totalTransactions' | 'totalPairs'>
  )> }
);